+++ src/chain/mocks.ts
@@ -1,0 +1,151 @@
+export const delegatesMock = [
+  {
+    account: {
+      name: 'hamid',
+      address: '128iGaUF4zV1L2bXtAH9UmocvqpSLrJFL5TxZYZkjoSf16Hv',
+    },
+    roles: ['nominatoe', 'fellow', 'validator'],
+    bio: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Proin vel facilisis augue. Praesent nec imperdiet diam. Etiam tempus convallis sollicitudin. Read more',
+    stats: [
+      { title: 'Most active track', value: 'Staking admin' },
+      { title: 'Participation', value: '75.71%' },
+      { title: 'Representing', value: '135 addresses' },
+      { title: 'DOT', value: '24k DOT' },
+    ],
+  },
+  {
+    account: {
+      name: 'hamid',
+      address: '128iGaUF4zV1L2bXtAH9UmocvqpSLrJFL5TxZYZkjoSf16Hv',
+    },
+    roles: ['nominatoe', 'fellow', 'validator'],
+    bio: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Proin vel facilisis augue. Praesent nec imperdiet diam. Etiam tempus convallis sollicitudin. Read more',
+    stats: [
+      { title: 'Most active track', value: 'Staking admin' },
+      { title: 'Participation', value: '75.71%' },
+      { title: 'Representing', value: '135 addresses' },
+      { title: 'DOT', value: '24k DOT' },
+    ],
+  },
+  {
+    account: {
+      name: 'hamid',
+      address: '128iGaUF4zV1L2bXtAH9UmocvqpSLrJFL5TxZYZkjoSf16Hv',
+    },
+    roles: ['nominatoe', 'fellow', 'validator'],
+    bio: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Proin vel facilisis augue. Praesent nec imperdiet diam. Etiam tempus convallis sollicitudin. Read more',
+    stats: [
+      { title: 'Most active track', value: 'Staking admin' },
+      { title: 'Participation', value: '75.71%' },
+      { title: 'Representing', value: '135 addresses' },
+      { title: 'DOT', value: '24k DOT' },
+    ],
+  },
+  {
+    account: {
+      name: 'hamid',
+      address: '128iGaUF4zV1L2bXtAH9UmocvqpSLrJFL5TxZYZkjoSf16Hv',
+    },
+    roles: ['nominatoe', 'fellow', 'validator'],
+    bio: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Proin vel facilisis augue. Praesent nec imperdiet diam. Etiam tempus convallis sollicitudin. Read more',
+    stats: [
+      { title: 'Most active track', value: 'Staking admin' },
+      { title: 'Participation', value: '75.71%' },
+      { title: 'Representing', value: '135 addresses' },
+      { title: 'DOT', value: '24k DOT' },
+    ],
+  },
+];
+
+export const tracksMock = [
+  {
+    title: 'Admin',
+    subtracks: [
+      {
+        title: 'Root',
+        description: 'Root is the track for the most important things',
+      },
+      {
+        title: 'Auction Admin',
+        description:
+          'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Mauris quis dui vitae leo aliquam iaculis. Maecenas faucibus nisi id lacus lacinia ornare. Phasellus in eros dignissim',
+      },
+      {
+        title: 'Staking Admin',
+        description:
+          'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Mauris quis dui vitae leo aliquam iaculis. Maecenas faucibus nisi id lacus lacinia ornare. Phasellus in eros dignissim',
+      },
+    ],
+  },
+  {
+    title: 'Governance',
+    subtracks: [
+      {
+        title: 'Lease Admin',
+        description: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit',
+      },
+      {
+        title: 'General Admin',
+        description: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit.',
+      },
+      {
+        title: 'Referendum Canceler',
+        description: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit.',
+      },
+      {
+        title: 'Referendum Killer',
+        description: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit.',
+      },
+    ],
+  },
+  {
+    title: 'Treasury',
+    subtracks: [
+      {
+        title: 'Small Tipper',
+        description: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit.',
+      },
+      {
+        title: 'Big Tipper',
+        description: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit.',
+      },
+      {
+        title: 'Small Spender',
+        description: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit.',
+      },
+      {
+        title: 'Medium Spender',
+        description: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit.',
+      },
+      {
+        title: 'Big Spender',
+        description: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit.',
+      },
+      {
+        title: 'Treasurer',
+        description: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit.',
+      },
+    ],
+  },
+  {
+    title: 'Fellowship',
+    subtracks: [
+      {
+        title: 'Whitelisted Caller',
+        description: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit.',
+      },
+      {
+        title: 'Members',
+        description: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit.',
+      },
+      {
+        title: 'Member Referenda',
+        description: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit.',
+      },
+      {
+        title: 'Fellowship Admin',
+        description: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit.',
+      },
+    ],
+  },
+];
