+++ src/ui/components/DelegateModal/index.tsx
@@ -1,0 +1,11 @@
+import { Modal } from '../../lib';
+import { TrackSelect } from '../TrackSelect';
+import { tracksMock } from '../../../chain/mocks';
+
+export function DelegateModal({ open, onClose }) {
+  return (
+    <Modal size="xl" open={open} onClose={onClose}>
+      <TrackSelect tracks={tracksMock} />
+    </Modal>
+  );
+}
