+++ src/ui/lib/Modal.tsx
@@ -20,7 +20,7 @@ const Modal = ({ className, children, size, open, onClose }: ModalProps) => {
     <div
       tabIndex={-1}
       className={`fixed top-0 left-0 right-0 z-50  flex h-full w-full overflow-y-auto overflow-x-hidden p-4 backdrop-blur  md:inset-0 ${display}`}
-      onClick={() => onClose()}
+      onClick={onClose && (() => onClose())}
     >
       <div
         className={`relative  mx-auto max-w-full self-center rounded-lg bg-white p-2 shadow ${
